+++ test/parallel/test-fs-watch-close-when-destroyed.js
@@ -5,6 +5,13 @@
 
 const common = require('../common');
 
+// fs-watch on folders have limited capability in AIX.
+// The testcase makes use of folder watching, and causes
+// hang. This behavior is documented. Skip this for AIX.
+
+if (common.isAIX)
+  common.skip('folder watch capability is limited in AIX.');
+
 if (common.isIBMi)
   common.skip('IBMi does not support `fs.watch()`');
