+++ test/parallel/test-fs-watch-recursive-promise.js
@@ -8,6 +8,13 @@ if (!common.hasCrypto)
 if (common.isIBMi)
   common.skip('IBMi does not support `fs.watch()`');
 
+// fs-watch on folders have limited capability in AIX.
+// The testcase makes use of folder watching, and causes
+// hang. This behavior is documented. Skip this for AIX.
+
+if (common.isAIX)
+  common.skip('folder watch capability is limited in AIX.');
+
 const { randomUUID } = require('crypto');
 const assert = require('assert');
 const path = require('path');
